Bei der Löschung eines Shop oder einer Kategorie werden alle Artikeln von dem Typ auf 1 gesetzt

+++ src/app/Einkauf/Category/Repository/DeleteSettingsRepository.php
@@ -9,7 +9,12 @@ class DeleteSettingsRepository
 
     public function deleteById($id): bool
     {
-        $db = new Database('EinkaufCat');
-        return $db->delete()->where("id", "=",":id")->args([':id' => $id])->run();
+        $db = new Database('EinkaufArticle');
+        if ($db->update(['cat_id'])->where("cat_id", "=", ":id")->args([':cat_id' => 1, ':id' => $id])->run()) {
+            $db = new Database('EinkaufCat');
+            return $db->delete()->where("id", "=", ":id")->args([':id' => $id])->run();
+        } else {
+            return false;
+        }
     }
 }
