+++ src/app/Einkauf/Shop/Repository/DeleteSettingsRepository.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Einkauf\Shop\Repository;
+
 use Foundation\Database\Database;
 
 class DeleteSettingsRepository
@@ -8,7 +9,12 @@ class DeleteSettingsRepository
 
     public function deleteById($id): bool
     {
-        $db = new Database('EinkaufShop');
-        return $db->delete()->where("id", "=",":id")->args([':id' => $id])->run();
+        $db = new Database('EinkaufArticle');
+        if ($db->update(['shop_id'])->where("shop_id", "=", ":id")->args([':shop_id' => 1, ':id' => $id])->run()) {
+            $db = new Database('EinkaufShop');
+            return $db->delete()->where("id", "=", ":id")->args([':id' => $id])->run();
+        } else {
+            return false;
+        }
     }
 }
